Use babel and webpack for theme.js
