+++ package.json
@@ -34,11 +34,14 @@
   },
   "devDependencies": {
     "@angular/compiler-cli": "^11.2.11",
+    "@babel/core": "^7.15.5",
+    "@babel/preset-env": "^7.15.6",
     "@ngtools/webpack": "^11.2.10",
     "@types/jquery": "^3.5.5",
     "@types/node": "^14.17.2",
     "@types/webcrypto": "^0.0.28",
     "@types/webpack": "^4.4.27",
+    "babel-loader": "^8.2.2",
     "clean-webpack-plugin": "^3.0.0",
     "copy-webpack-plugin": "^6.4.0",
     "cross-env": "^7.0.3",
@@ -47,6 +50,7 @@
     "file-loader": "^6.2.0",
     "gh-pages": "^3.1.0",
     "html-loader": "^1.3.2",
+    "html-webpack-injector": "1.1.4",
     "html-webpack-plugin": "^4.5.1",
     "mini-css-extract-plugin": "^1.5.0",
     "sass": "^1.32.10",
