+++ src/index.html
@@ -8,9 +8,6 @@
 
     <title page-title>Bitwarden Web Vault</title>
 
-    <script src="scripts/theme.js" crossorigin="anonymous" type="application/javascript"
-        integrity="sha384-qSFc8wSe0pmQX72GA/C7DD2EGxP6GCO8V7EgT6/MkpDqyk88CvlzBWQCzkgzQxSD"></script>
-
     <link rel="apple-touch-icon" sizes="180x180" href="images/icons/apple-touch-icon.png">
     <link rel="icon" type="image/png" sizes="32x32" href="images/icons/favicon-32x32.png">
     <link rel="icon" type="image/png" sizes="16x16" href="images/icons/favicon-16x16.png">
