+++ webpack.config.js
@@ -3,6 +3,7 @@ const fs = require('fs');
 const webpack = require('webpack');
 const { CleanWebpackPlugin } = require('clean-webpack-plugin');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
+const HtmlWebpackInjector = require('html-webpack-injector');
 const CopyWebpackPlugin = require('copy-webpack-plugin');
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const TerserPlugin = require('terser-webpack-plugin');
@@ -51,10 +52,11 @@ const moduleRules = [
     {
         test: /theme\.js$/i,
         use: [{
-            loader: 'file-loader',
+            loader: 'babel-loader',
             options: {
-                name: '[name].[ext]',
-                outputPath: 'scripts/',
+                presets: [
+                    ['@babel/preset-env', { targets: "defaults" }]
+                ]
             },
         }],
     },
@@ -90,8 +92,9 @@ const plugins = [
     new HtmlWebpackPlugin({
         template: './src/index.html',
         filename: 'index.html',
-        chunks: ['app/polyfills', 'app/vendor', 'app/main'],
+        chunks: ['theme_head', 'app/polyfills', 'app/vendor', 'app/main'],
     }),
+    new HtmlWebpackInjector(),
     new HtmlWebpackPlugin({
         template: './src/connectors/duo.html',
         filename: 'duo-connector.html',
@@ -232,6 +235,7 @@ const webpackConfig = {
         'connectors/duo': './src/connectors/duo.ts',
         'connectors/sso': './src/connectors/sso.ts',
         'connectors/captcha': './src/connectors/captcha.ts',
+        'theme_head': './src/theme.js',
     },
     externals: {
         'u2f': 'u2f',
